+++ packages/HitTriage/src/app/accordeons/new-template-accordeon.ts
@@ -8,9 +8,11 @@ import * as C from '../consts';
 import '../../../css/hit-triage.css';
 import {chemFunctionsDialog} from '../dialogs/functions-dialog';
 import {ItemType, ItemsGrid} from '@datagrok-libraries/utils/src/items-grid';
+import {HitAppBase} from '../hit-app-base';
 
 
-export async function createTemplateAccordeon(dataSourceFunctionMap: { [key: string]: DG.Func | DG.DataQuery },
+export async function createTemplateAccordeon(app: HitAppBase<any>,
+  dataSourceFunctionMap: { [key: string]: DG.Func | DG.DataQuery },
 ): Promise<INewTemplateResult<HitTriageTemplate>> {
   const functions = DG.Func.find({tags: [C.HitTriageComputeFunctionTag]});
   const availableTemplates = (await _package.files.list('Hit Triage/templates')).map((file) => file.name.slice(0, -5));
@@ -77,7 +79,7 @@ export async function createTemplateAccordeon(dataSourceFunctionMap: { [key: str
       functions: [],
     },
   } as unknown as HitTriageTemplate;
-  const funcInput = await chemFunctionsDialog((res) => {funcDialogRes = res;}, () => null,
+  const funcInput = await chemFunctionsDialog(app, (res) => {funcDialogRes = res;}, () => null,
     dummyTemplate, false);
   funcInput.root.classList.add('hit-triage-new-template-functions-input');
   if (Object.entries(dataSourceFunctionMap).length === 0) {