+++ packages/HitTriage/src/app/hit-triage-app.ts
@@ -1,21 +1,21 @@
 import * as grok from 'datagrok-api/grok';
 import * as ui from 'datagrok-api/ui';
 import * as DG from 'datagrok-api/dg';
-import {HitTriageCampaign, IComputeDialogResult, IFunctionArgs,
+import {HitTriageCampaign, IFunctionArgs,
   HitTriageTemplate, HitTriageTemplateIngest, IngestType, HitTriageCampaignStatus,
   HitDesignTemplate} from './types';
 import {InfoView} from './hit-triage-views/info-view';
 import {SubmitView} from './hit-triage-views/submit-view';
 import {CampaignIdKey, CampaignJsonName, CampaignTableName,
   HTQueryPrefix, HTScriptPrefix, HitSelectionColName, i18n} from './consts';
-import {addBreadCrumbsToRibbons, checkRibbonsHaveSubmit,
-  joinQueryResults, modifyUrl, toFormatedDateString} from './utils';
+import {addBreadCrumbsToRibbons, checkRibbonsHaveSubmit, modifyUrl, toFormatedDateString} from './utils';
 import {_package} from '../package';
 import '../../css/hit-triage.css';
 import {chemFunctionsDialog} from './dialogs/functions-dialog';
 import {HitAppBase} from './hit-app-base';
 import {HitBaseView} from './base-view';
 import {saveCampaignDialog} from './dialogs/save-campaign-dialog';
+import {calculateColumns} from './utils/calculate-single-cell';
 export class HitTriageApp extends HitAppBase<HitTriageTemplate> {
   multiView: DG.MultiView;
 
@@ -111,10 +111,10 @@ export class HitTriageApp extends HitAppBase<HitTriageTemplate> {
         queries[fName] = query.args;
       });
 
-      await this.calculateColumns({
+      await calculateColumns({
         descriptors: template.compute.descriptors.enabled ? template.compute.descriptors.args : [],
         externals: funcs, scripts, queries,
-      });
+      }, this.dataFrame!, this._molColName!);
     };
     const curView = grok.shell.v;
     const pickV = grok.shell.addView(this.pickView);
@@ -152,71 +152,6 @@ export class HitTriageApp extends HitAppBase<HitTriageTemplate> {
 
   set campaign(campaign: HitTriageCampaign | undefined) {this._campaign = campaign;}
 
-  public async calculateColumns(resultMap: IComputeDialogResult, view?: DG.TableView) {
-    const previousColumns = this.dataFrame!.columns.names();
-    if (resultMap.descriptors && resultMap.descriptors.length > 0)
-      await grok.chem.descriptors(this.dataFrame!, this.molColName!, resultMap.descriptors);
-
-    for (const funcName of Object.keys(resultMap.externals)) {
-      const props = resultMap.externals[funcName];
-      const f = DG.Func.find({package: funcName.split(':')[0], name: funcName.split(':')[1]})[0];
-      const tablePropName = f.inputs[0].name;
-      const colPropName = f.inputs[1].name;
-      if (props)
-        await f.apply({...props, [tablePropName]: this.dataFrame!, [colPropName]: this.molColName});
-    };
-    for (const scriptName of Object.keys(resultMap.scripts ?? {})) {
-      const props = resultMap.scripts![scriptName];
-      if (props) {
-        // props['table'] = this.dataFrame!;
-        // props['molecules'] = this.molColName!;
-        const scriptParts = scriptName.split(':');
-        const scriptId = scriptParts[2];
-        if (!scriptId)
-          continue;
-        try {
-          const s = await grok.dapi.scripts.find(scriptId);
-          if (!s)
-            continue;
-          const tablePropName = s.inputs[0].name;
-          const colPropName = s.inputs[1].name;
-          await s.apply({...props, [tablePropName]: this.dataFrame!, [colPropName]: this.molColName});
-        } catch (e) {
-          console.error(e);
-        }
-      }
-    };
-
-    // handling queries
-    for (const queryName of Object.keys(resultMap.queries ?? {})) {
-      const props = resultMap.queries![queryName];
-      if (props) {
-        const queryParts = queryName.split(':');
-        const queryId = queryParts[2];
-        if (!queryId)
-          continue;
-        try {
-          const s = await grok.dapi.queries.find(queryId);
-          if (!s)
-            continue;
-          const listPropName = s.inputs[0].name;
-          const molList = this.dataFrame!.col(this.molColName!)!.toList();
-          const resDf: DG.DataFrame = await s.apply({...props, [listPropName]: molList});
-          if (resDf)
-            await joinQueryResults(this.dataFrame!, this.molColName!, resDf);
-        } catch (e) {
-          console.error(e);
-        }
-      }
-    };
-
-    if (view) {
-      const filterGroup = view.getFiltersGroup();
-    this.dataFrame!.columns.names().filter((colName) => !previousColumns.includes(colName))
-      .forEach((colName) => {filterGroup.add({type: this.getFilterType(colName), column: colName});});
-    }
-  }
-
   /**
    * A view that lets you filter the molecules using either molecules, or
    * their properties derived at the enrichment step.
@@ -226,7 +161,7 @@ export class HitTriageApp extends HitAppBase<HitTriageTemplate> {
     let resolvePromise: Function | null = null;
     this._pickViewPromise = new Promise<void>((res) => {resolvePromise = res;});
     const getComputeDialog = async () => {
-      chemFunctionsDialog(async (resultMap) => {
+      chemFunctionsDialog(this, async (resultMap) => {
         const oldDescriptors = this.template!.compute.descriptors.args;
         const oldFunctions = this.template!.compute.functions;
         const oldScripts = this.template!.compute.scripts ?? [];
@@ -302,10 +237,10 @@ export class HitTriageApp extends HitAppBase<HitTriageTemplate> {
           const fName = `${HTQueryPrefix}:${query.name}:${query.id}`;
           externalQueries[fName] = query.args;
         });
-        await this.calculateColumns(
+        await calculateColumns(
           {descriptors: uncalculatedDescriptors, externals: externalFuncs,
             scripts: externalScripts, queries: externalQueries},
-          view);
+          this.dataFrame!, this._molColName!);
         this.saveCampaign('In Progress', false);
       }, () => null, this.template!, true);
     };